REF: Use new write_tiff everywhere
Other direct calls to tiffile.imsave can be replaced by the new
write_tiff; this reduced code length slightly.
